# api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MediSync - Healthcare Resource Scheduling API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Doctor {
  id: number;
  name: string;
  specialization: string;
}

export interface DoctorInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  specialization: string;
}

export interface Appointment {
  id: number;
  patientName: string;
  doctorId: number;
  startTime: string;
}

export interface AppointmentWithDoctor {
  id: number;
  patientName: string;
  doctorId: number;
  startTime: string;
  doctor: Doctor;
}

export interface AppointmentInput {
  /** @minLength 1 */
  patientName: string;
  doctorId: number;
  startTime: string;
}

export interface DoctorAppointmentCount {
  doctorId: number;
  doctorName: string;
  specialization: string;
  count: number;
}

export interface AppointmentSummary {
  total: number;
  upcoming: number;
  past: number;
  byDoctor: DoctorAppointmentCount[];
}

export interface ErrorResponse {
  error: string;
}