# appointmentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MediSync - Healthcare Resource Scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentInput {
  /** @minLength 1 */
  patientName: string;
  doctorId: number;
  startTime: Date;
}