# errorResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MediSync - Healthcare Resource Scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
}