# openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MediSync - Healthcare Resource Scheduling API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List upcoming appointments
      description: Returns upcoming appointments with doctor details
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppointmentWithDoctor"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      description: Books an appointment. Returns 409 if the doctor is already booked at the requested time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentWithDoctor"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Scheduling conflict — doctor already booked at that time
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /appointments/{id}:
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Appointment cancelled
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialization:
          type: string
      required:
        - id
        - name
        - specialization
...
    AppointmentWithDoctor:
      type: object
      properties:
        id:
          type: integer
        patientName:
          type: string
        doctorId:
          type: integer
        startTime:
          type: string
          format: date-time
        doctor:
          $ref: "#/components/schemas/Doctor"
      required:
        - id
        - patientName
        - doctorId
        - startTime
        - doctor
    AppointmentInput:
      type: object
      properties:
        patientName:
          type: string
          minLength: 1
        doctorId:
          type: integer
        startTime:
          type: string
          format: date-time
      required:
        - patientName
        - doctorId
        - startTime
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error